requirements.txt
packaging
sentencepiece
smplx
peft==0.4.0
einops==0.4.1
fastapi==0.100.1
gradio==3.39.0
markdown2==2.4.10
numpy==1.24.2
openai==0.27.8
opencv_python==4.8.0.74
Pillow==9.4.0
pycocotools==2.0.6
ray==2.6.1
Requests==2.31.0
shortuuid==1.0.11
tqdm==4.64.1
transformers==4.31.0
uvicorn==0.23.2
scipy==1.11.2
bitsandbytes==0.41.1
tensorboard==2.17.0
scikit-learn==1.5.1
scikit-image==0.24.0
trimesh
torchshow
pillow_heif
git+https://github.com/facebookresearch/pytorch3d.git@stable
timm
git+https://github.com/vguzov/videoio
pyrender
wandb